CRAB-45697: In Development Watch (#8)
The `python addon.py watch` command now allows non-admin users to watch deployed "In Development" Add-ons.

+++ addon_template/_dev_tools/addon_tasks/watch.py
@@ -35,5 +35,6 @@ def watch(args):
         except KeyboardInterrupt:
             print('Stopping watch')
             for process in processes.values():
-                process.terminate()
+                if process is not None: # Elements that pass on the watch command will return None
+                    process.terminate()
             break
